Resolved issue #69
- Added functionality for the 'I am the FC button'
Pressing this button will change the listed FC, and the object used by the waitlist for fleet ESI queries. The waitlist no longer needs to be shut down and opened under the new FC.

- Added functionality to the backseat button.
When this button is pressed, the user is set as the backseat, if the backseat user presses the button they will be unset. If the FC presses this button, the backseat will also be unset.