Added a waitlist banner.
## Waitlist Banner
Allows FCs to set a global message for all pilots to see on the waitlist `/` page. FCs can set two types an info (blue) and a warning (red) banner.

Only one banner can be displayed at a time, setting a new banner will hide the old one.  Any FC can set or hide a banner. To hide a banner click on the __admin remove__ link on the banner on the `/` index page.

---
### Changes
- Added a `waitlistBanner.js` file with methods for hiding, creating and getting banner objects.
- Added an html snippet in `waitlist.njk` at the top of the page that will only load if there is a bannerObject.
- Added a banner menu option and modal to create new banners.
- Added new `/internal-api/` routes as well as methods in the API controller to handle banner requests.
- Added css for banner alerts in `custom.css`.
- Added a function in `pagesController` to get the last banner and pass it to the view

#### Non Related Changes
- Removed ```log.debug(`User removed themselves from waitlist: ${characterID}`);``` from `globalWaitlist.js` as this is now stored in the database and rendered in a view.

### Known Issues
_N/A at this time_.