Users object cleanup
Users object has been restructured. The role fields are now in an object and every view and controller has been updated.

Corporation and alliances are in a clean object holding only the id and name. This is now gained from the cache.

The users method `getUserDataFromID` has been renamed to `getPilotAffiliation` and the code has been reduced. Only one ESI call now happens in here.